# docs/swagger.yaml
basePath: /v1
definitions:
  models.Refresh:
    properties:
      refresh:
        type: string
    type: object
  models.Token:
    properties:
      refreshtoken:
        type: string
      status:
        type: string
      token:
        type: string
    type: object
  models.User:
    properties:
      id:
        type: string
      password:
        type: string
      refresh:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  models.UserToken:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: User API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Create a new access token.
      parameters:
      - description: Create Token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.UserToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Token'
            type: array
      summary: create a new access token
      tags:
      - Token
  /refresh:
    post:
      consumes:
      - application/json
      description: Create a new token.
      parameters:
      - description: Create Token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.Refresh'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Token'
            type: array
      summary: create a new token
      tags:
      - Token
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user.
      parameters:
      - description: Signup User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: create a new user
      tags:
      - User
  /user:
    get:
      consumes:
      - application/json
      description: Get user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get user
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"